Added scheduling functionality. Renamed to “django-scheduled-mailer”.

+++ scheduled_mailer/engine.py
@@ -1,6 +1,7 @@
 import time
 import smtplib
 import logging
+import datetime
 
 from lockfile import FileLock, AlreadyLocked, LockTimeout
 from socket import error as socket_error
@@ -15,7 +16,7 @@ except ImportError:
     from django.core.mail import SMTPConnection
     get_connection = lambda backend=None, fail_silently=False, **kwds: SMTPConnection(fail_silently=fail_silently)
 
-from mailer.models import Message, DontSendEntry, MessageLog
+from scheduled_mailer.models import Message, DontSendEntry, MessageLog
 
 
 # when queue is empty, how long to wait (in seconds) before checking again
@@ -74,23 +75,26 @@ def send_all():
     try:
         connection = None
         for message in prioritize():
-            try:
-                if connection is None:
-                    connection = get_connection(backend=EMAIL_BACKEND)
-                logging.info("sending message '%s' to %s" % (message.subject.encode("utf-8"), u", ".join(message.to_addresses).encode("utf-8")))
-                email = message.email
-                email.connection = connection
-                email.send()
-                MessageLog.objects.log(message, 1) # @@@ avoid using literal result code
-                message.delete()
-                sent += 1
-            except (socket_error, smtplib.SMTPSenderRefused, smtplib.SMTPRecipientsRefused, smtplib.SMTPAuthenticationError), err:
-                message.defer()
-                logging.info("message deferred due to failure: %s" % err)
-                MessageLog.objects.log(message, 3, log_message=str(err)) # @@@ avoid using literal result code
-                deferred += 1
-                # Get new connection, it case the connection itself has an error.
-                connection = None
+            if message.when_send > datetime.datetime.now():
+                continue
+            else:
+                try:
+                    if connection is None:
+                        connection = get_connection(backend=EMAIL_BACKEND)
+                    logging.info("sending message '%s' to %s" % (message.subject.encode("utf-8"), u", ".join(message.to_addresses).encode("utf-8")))
+                    email = message.email
+                    email.connection = connection
+                    email.send()
+                    MessageLog.objects.log(message, 1) # @@@ avoid using literal result code
+                    message.delete()
+                    sent += 1
+                except (socket_error, smtplib.SMTPSenderRefused, smtplib.SMTPRecipientsRefused, smtplib.SMTPAuthenticationError), err:
+                    message.defer()
+                    logging.info("message deferred due to failure: %s" % err)
+                    MessageLog.objects.log(message, 3, log_message=str(err)) # @@@ avoid using literal result code
+                    deferred += 1
+                    # Get new connection, it case the connection itself has an error.
+                    connection = None
     finally:
         logging.debug("releasing lock...")
         lock.release()
